Added StorageCyclicBoundary Constraint
Minor modifications to other files were made
