+++ hypatia/backend/constraints/ConstraintList.py
@@ -6,12 +6,12 @@ from hypatia.backend.constraints.TotalCapacityRegional import TotalCapacityRegio
 from hypatia.backend.constraints.NewCapacityRegional import NewCapacityRegional
 from hypatia.backend.constraints.TechEfficency import TechEfficency
 from hypatia.backend.constraints.AnnualProductionRegional import AnnualProductionRegional
-#Electricita esplisita
 from hypatia.backend.constraints.EmissionCapRegional import EmissionCapRegional
 from hypatia.backend.constraints.EmissionCapGlobal import EmissionCapGlobal
 from hypatia.backend.constraints.EmissionConsumedRegional import EmissionConsumedRegional
 from hypatia.backend.constraints.StorageMaxMinChange import StorageMaxMinChange
 from hypatia.backend.constraints.StorageMaxFlowInOut import StorageMaxFlowInOut
+from hypatia.backend.constraints.StorageCyclicBoundary import StorageCyclicBoundary
 from hypatia.backend.constraints.LineTotalCapacity import LineTotalCapacity
 from hypatia.backend.constraints.TotalCapacityGlobal import TotalCapacityGlobal
 from hypatia.backend.constraints.NewCapacityGlobal import NewCapacityGlobal
@@ -20,12 +20,11 @@ from hypatia.backend.constraints.LineAvailability import LineAvailability
 from hypatia.backend.constraints.LineNewCapacity import LineNewCapacity
 from hypatia.backend.constraints.LandUsageRegional import LandUsageRegional
 from hypatia.backend.constraints.LandUsageGlobal import LandUsageGlobal
-# from hypatia.backend.constraints.ElectrolysisConsumption import ElectrolysisConsumption
-# da modificare il parte di eletricita esplisita
+# from hypatia.backend.constraints.ElectrolysisConsumption import ElectrolysisConsumption # To make it work, "Electricity" carrier must be explicitly specified
 from hypatia.backend.constraints.RenewableProductionRegional import RenewableProductionRegional
 from hypatia.backend.constraints.RenewableProductionGlobal import RenewableProductionGlobal  
 #from hypatia.backend.constraints.NewCapacityNonneg import NewCapacityNonneg
-#from hypatia.backend.constraints.ProductionAndStorageUseRamping import ProductionAndStorageUseRamping
+#from hypatia.backend.constraints.ProductionRamp import ProductionRamp
 
 CONSTRAINTS = [
     Balance,
@@ -41,6 +40,7 @@ CONSTRAINTS = [
     EmissionConsumedRegional,
     StorageMaxMinChange,
     StorageMaxFlowInOut,
+    StorageCyclicBoundary,
     LineTotalCapacity,
     TotalCapacityGlobal,
     NewCapacityGlobal,
@@ -52,6 +52,7 @@ CONSTRAINTS = [
     RenewableProductionRegional,
     LandUsageRegional,
     LandUsageGlobal,
+    # ProductionRamp,
     # ElectrolysisConsumption
 ]
