+++ hypatia/backend/constraints/StorageMaxMinChange.py
@@ -12,34 +12,38 @@ class StorageMaxMinChange(Constraint):
         assert hasattr(self.variables, 'totalcapacity'), "totalcapacity must be defined"
 
     def rules(self):
+        timeslice_fraction = self.model_data.settings.timeslice_fraction
+        if not isinstance(timeslice_fraction, int):
+            timeslice_fraction.shape = (len(self.model_data.settings.time_steps), 1)
+        print("timeslice fraction shape:", timeslice_fraction.shape)    
         rules = []
         for reg in get_regions_with_storage(self.model_data.settings):
             for indx, year in enumerate(self.model_data.settings.years):
+                print("storage capacity shape:", timeslice_fraction.shape)
+                annual_storage_capacity = cp.multiply(
+                    self.variables.totalcapacity[reg]["Storage"][indx : indx + 1, :],
+                    timeslice_fraction) * 8760 # shape = (ts, techs)
                 
                 rules.append(
-                    self.variables.totalcapacity[reg]["Storage"][indx : indx + 1, :]
+                    annual_storage_capacity
                     - self.variables.storage_SOC[reg][
-                        indx
-                        * len(self.model_data.settings.time_steps) : (indx + 1)
-                        * len(self.model_data.settings.time_steps),
-                        :,
-                    ]
+                        indx* len(self.model_data.settings.time_steps)
+                        : (indx + 1) * len(self.model_data.settings.time_steps),
+                        :] # shape = (ts, techs)
                     >= 0
-                )
+                )                
                 
                 rules.append(
                     self.variables.storage_SOC[reg][
-                        indx
-                        * len(self.model_data.settings.time_steps) : (indx + 1)
-                        * len(self.model_data.settings.time_steps),
-                        :,
-                    ]
+                        indx * len(self.model_data.settings.time_steps) 
+                        : (indx + 1) * len(self.model_data.settings.time_steps),
+                        :] # shape = (ts, techs)
                     - cp.multiply(
-                        self.variables.totalcapacity[reg]["Storage"][indx : indx + 1, :],
+                        annual_storage_capacity,
                         self.model_data.regional_parameters[reg]["storage_min_SOC"].values[indx : indx + 1, :]
-                    )
+                    ) # shape = (ts, techs)
                     >= 0
                 )
-                
+                      
         return rules
